Defining main loop/menu functionality.
Defining creature generator for the player character and on-par enemies.
(A separate function will handle the creation of "grunt" units.)

Per my advisor's recommendation, I'll be terming the 15-battle design element in lieu of a proof of concept that shows the ability to transfer the player's party between rounds of combat.
(Still has to be implemented.)